+++ pypinyin/constants.py
@@ -27,7 +27,9 @@ re_phonetic_symbol_source = ''.join(PHONETIC_SYMBOL.keys())
 # 匹配带声调字符的正则表达式
 RE_PHONETIC_SYMBOL = r'[' + re.escape(re_phonetic_symbol_source) + r']'
 # 匹配使用数字标识声调的字符的正则表达式
-RE_TONE2 = r'([aeoiuvnm])([0-4])$'
+RE_TONE2 = r'([aeoiuvnm])([1-4])$'
+# 匹配 TONE2 中标识韵母声调的正则表达式
+RE_TONE3 = re.compile('^([a-z]+)([1-4])([a-z]*)$')
 # 有拼音的汉字
 if SUPPORT_UCS4:
     RE_HANS = re.compile(
@@ -54,19 +56,25 @@ else:
 PINYIN_STYLE = {
     'NORMAL': 0,          # 普通风格，不带声调
     'TONE': 1,            # 标准风格，声调在韵母的第一个字母上
-    'TONE2': 2,           # 声调在拼音之后，使用数字 1~4 标识
+    'TONE2': 2,           # 声调在韵母之后，使用数字 1~4 标识
+    'TONE3': 8,           # 声调在拼音之后，使用数字 1~4 标识
     'INITIALS': 3,        # 仅保留声母部分
     'FIRST_LETTER': 4,    # 仅保留首字母
     'FINALS': 5,          # 仅保留韵母部分，不带声调
     'FINALS_TONE': 6,     # 仅保留韵母部分，带声调
-    'FINALS_TONE2': 7,    # 仅保留韵母部分，声调在拼音之后，使用数字 1~4 标识
+    'FINALS_TONE2': 7,    # 仅保留韵母部分，声调在韵母之后，使用数字 1~4 标识
+    'FINALS_TONE3': 9,    # 仅保留韵母部分，声调在拼音之后，使用数字 1~4 标识
+    'BOPOMOFO': 10,       # 注音符号，带声调，阴平（第一声）不标
+    'BOPOMOFO_FIRST': 11,  # 注音符号首字母
 }
 # 普通风格，不带声调
 NORMAL = STYLE_NORMAL = PINYIN_STYLE['NORMAL']
 # 标准风格，声调在韵母的第一个字母上
 TONE = STYLE_TONE = PINYIN_STYLE['TONE']
-# 声调在拼音之后，使用数字 1~4 标识
+# 声调在韵母之后，使用数字 1~4 标识
 TONE2 = STYLE_TONE2 = PINYIN_STYLE['TONE2']
+# 声调在拼音之后，使用数字 1~4 标识
+TONE3 = STYLE_TONE3 = PINYIN_STYLE['TONE3']
 # 仅保留声母部分
 INITIALS = STYLE_INITIALS = PINYIN_STYLE['INITIALS']
 # 仅保留首字母
@@ -75,8 +83,14 @@ FIRST_LETTER = STYLE_FIRST_LETTER = PINYIN_STYLE['FIRST_LETTER']
 FINALS = STYLE_FINALS = PINYIN_STYLE['FINALS']
 # 仅保留韵母部分，带声调
 FINALS_TONE = STYLE_FINALS_TONE = PINYIN_STYLE['FINALS_TONE']
-# 仅保留韵母部分，声调在拼音之后，使用数字 1~4 标识
+# 仅保留韵母部分，声调在韵母之后，使用数字 1~4 标识
 FINALS_TONE2 = STYLE_FINALS_TONE2 = PINYIN_STYLE['FINALS_TONE2']
+# 仅保留韵母部分，声调在拼音之后，使用数字 1~4 标识
+FINALS_TONE3 = STYLE_FINALS_TONE3 = PINYIN_STYLE['FINALS_TONE3']
+# 注音符号，带声调，阴平（第一声）不标
+BOPOMOFO = STYLE_BOPOMOFO = PINYIN_STYLE['BOPOMOFO']
+# 注音符号首字母
+BOPOMOFO_FIRST = STYLE_BOPOMOFO_FIRST = PINYIN_STYLE['BOPOMOFO_FIRST']
 
 U_FINALS_EXCEPTIONS_MAP = {
     u'ū': u'ǖ',
@@ -84,3 +98,38 @@ U_FINALS_EXCEPTIONS_MAP = {
     u'ǔ': u'ǚ',
     u'ù': u'ǜ',
 }
+
+BOPOMOFO_REPLACE = (
+    (re.compile('^m(\d)$'), 'mu\\1'),  # 呣
+    (re.compile('^n(\d)$'), 'N\\1'),  # 嗯
+    (re.compile('^r5$'), 'er5'),  # 〜兒
+    (re.compile('iu'), 'iou'),
+    (re.compile('ui'), 'uei'),
+    (re.compile('ong'), 'ung'),
+    (re.compile('^yi?'), 'i'),
+    (re.compile('^wu?'), 'u'),
+    (re.compile('iu'), 'v'),
+    (re.compile('^([jqx])u'), '\\1v'),
+    (re.compile('([iuv])n'), '\\1en'),
+    (re.compile('^zhi?'), 'Z'),
+    (re.compile('^chi?'), 'C'),
+    (re.compile('^shi?'), 'S'),
+    (re.compile('^([zcsr])i'), '\\1'),
+    (re.compile('ai'), 'A'),
+    (re.compile('ei'), 'I'),
+    (re.compile('ao'), 'O'),
+    (re.compile('ou'), 'U'),
+    (re.compile('ang'), 'K'),
+    (re.compile('eng'), 'G'),
+    (re.compile('an'), 'M'),
+    (re.compile('en'), 'N'),
+    (re.compile('er'), 'R'),
+    (re.compile('eh'), 'E'),
+    (re.compile('([iv])e'), '\\1E'),
+    (re.compile('([^0-4])$'), '\\g<1>0'),
+    (re.compile('1$'), ''),
+)
+BOPOMOFO_TABLE = dict(zip(
+    'bpmfdtnlgkhjqxZCSrzcsiuvaoeEAIOUMNKGR2340',
+    'ㄅㄆㄇㄈㄉㄊㄋㄌㄍㄎㄏㄐㄑㄒㄓㄔㄕㄖㄗㄘㄙㄧㄨㄩㄚㄛㄜㄝㄞㄟㄠㄡㄢㄣㄤㄥㄦˊˇˋ˙'
+))